シミュレーション設定の操作がredo、undoの対象になっている シミュレーションパラメータにファイル名「DefaultSets.csv」が読み込める Stepperの名前を変更するとRedoできない
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@3440 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/PathwayWindow/PathwayWindow/Animation/MovieAnimationItem.cs
@@ -318,8 +318,8 @@ namespace Ecell.IDE.Plugins.PathwayWindow.Animation
                 e.Cancel = true;
                 return;
             }
-            double dummy;
-            if (!double.TryParse(text, out dummy) || dummy <= 0)
+            int dummy;
+            if (!Int32.TryParse(text, out dummy) || dummy <= 0)
             {
                 Util.ShowErrorDialog(string.Format(MessageResources.ErrInvalidValue, maxSizeRadio.Text));
                 maxSizeTextBox.Text = Convert.ToString(_maxSize);
